loop rolling

+++ src/color/ft_colorlcode.c
@@ -14,7 +14,7 @@
 #include "ft_str.h"
 #include "ft_strarr.h"
 
-ssize_t		ft_colorlcode(const char *str)
+int		ft_colorlcode(const char *str)
 {
 	char	**colors;
 	size_t	i;
