+++ src/color/ft_colorncode.c
@@ -14,7 +14,7 @@
 #include "ft_str.h"
 #include "ft_strarr.h"
 
-ssize_t		ft_colorncode(const char *str, size_t n)
+int		ft_colorncode(const char *str, size_t n)
 {
 	char	**colors;
